requirements.txt
# Core
numpy>=1.22
pandas>=1.5
scikit-learn>=1.1
xgboost>=1.7

# Explainability
shap>=0.43
lime>=0.2.0.1
imbalanced-learn>=0.10

# UI & plots
gradio>=4.0,<5
matplotlib>=3.6
seaborn>=0.12

# Images
Pillow>=9.0